pyproject.toml
[tool.poetry]
name = "livraria-django-paulo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.0"
python-magic = "^0.4.27"
Pillow = "^9.2.0"
drf-spectacular = "^0.24.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
